+++ pkg/graphd/bitmapper/bitmapper.go
@@ -21,11 +21,26 @@ var ShardDBPattern = "shard_%d.db"
 
 var tracer = otel.Tracer("bitmapper")
 
-type Entity struct {
-	LK sync.RWMutex
-	BM *roaring.Bitmap
+// Bitmapper is a struct that manages collections of Roaring Bitmaps
+// Collections of bitmaps are organized into Groups that share UID mappings
+// Each Group has its own SQLite database for persisting the UID to string ID mappings
+// as well as a set of sharded SQLite databases for persisting the Roaring Bitmaps
+type Bitmapper struct {
+	groups  map[string]*Group
+	groupLk sync.RWMutex
+
+	CrossGroupLk sync.RWMutex
+
+	// Root directory for SQLite files, one subfolder per Group
+	dbDir string
 }
 
+// Group is a struct that manages a collection of Roaring Bitmaps
+// Bitmaps are organized by UID, with mappings to string IDs
+// Each Group has its own SQLite database for persisting the UID to string ID mappings
+// as well as a set of sharded SQLite databases for persisting the Roaring Bitmaps
+// Groups use an ARC cache for lazily loading Bitmaps from disk to keep memory usage manageable
+// Writes to Entities in a Group are immediately persisted to disk
 type Group struct {
 	name string
 
@@ -49,27 +64,27 @@ type Group struct {
 	shardSize uint32
 }
 
-type Bitmapper struct {
-	groups  map[string]*Group
-	groupLk sync.RWMutex
-
-	CrossGroupLk sync.RWMutex
-
-	// Root directory for SQLite files, one subfolder per group
-	dbDir string
+// Entity is a member of a Group
+// It contains a Roaring Bitmap and a lock for concurrent access
+// Entities should not be modified directly, but instead through the Group's UpdateEntity method
+type Entity struct {
+	LK sync.RWMutex
+	BM *roaring.Bitmap
 }
 
+// Config is the configuration for a Bitmapper instance
+type Config struct {
+	// Root directory for SQLite files, one subfolder per group
+	DBDir  string
+	Groups []GroupConfig
+}
+
+// GroupConfig is the configuration for a Group instance
 type GroupConfig struct {
 	Name      string
 	ShardSize uint32
 	CacheSize int
 	dbDir     string
-}
-
-type BitmapperConfig struct {
-	// Root directory for SQLite files, one subfolder per group
-	DBDir  string
-	Groups []GroupConfig
 }
 
 // NewGroup creates a new Group instance
@@ -117,7 +132,7 @@ func NewGroup(ctx context.Context, cfg GroupConfig) (*Group, error) {
 }
 
 // NewBitmapper creates a new Bitmapper instance and initializes the groups
-func NewBitmapper(ctx context.Context, cfg BitmapperConfig) (*Bitmapper, error) {
+func NewBitmapper(ctx context.Context, cfg Config) (*Bitmapper, error) {
 	ctx, span := tracer.Start(ctx, "NewBitmapper")
 	defer span.End()
 
@@ -186,6 +201,7 @@ func (b *Bitmapper) GetGroup(ctx context.Context, name string) (*Group, error) {
 	return group, nil
 }
 
+// PeekNextUID returns the next UID that will be assigned
 func (g *Group) PeekNextUID() uint32 {
 	g.nextLk.Lock()
 	defer g.nextLk.Unlock()
@@ -193,6 +209,11 @@ func (g *Group) PeekNextUID() uint32 {
 	uid := g.nextUID
 
 	return uid
+}
+
+// GetShardSize returns the shard size for the group
+func (g *Group) GetShardSize() uint32 {
+	return g.shardSize
 }
 
 // ErrorUIDNotFound is returned when a UID is not found
@@ -285,161 +306,9 @@ func (g *Group) UpdateEntity(ctx context.Context, uid uint32, bm *roaring.Bitmap
 	return nil
 }
 
-func (g *Group) loadBitmap(ctx context.Context, uid uint32) (*roaring.Bitmap, error) {
-	shard := int(uid / g.shardSize)
-	db, err := g.initShardDB(ctx, shard)
-	if err != nil {
-		return nil, fmt.Errorf("failed to initialize shard db: %w", err)
-	}
-
-	var bmBytes []byte
-	if err := db.QueryRowContext(ctx, `SELECT compressed_bitmap FROM entities WHERE uid = ?`, uid).Scan(&bmBytes); err != nil {
-		// If the entity doesn't exist, return an empty bitmap
-		if err == sql.ErrNoRows {
-			return roaring.NewBitmap(), nil
-		}
-		return nil, fmt.Errorf("failed to query entity: %w", err)
-	}
-
-	bm := roaring.New()
-	if _, err := bm.FromBuffer(bmBytes); err != nil {
-		return nil, fmt.Errorf("failed to unmarshal bitmap: %w", err)
-	}
-
-	return bm, nil
-}
-
-func (g *Group) updateBitmap(ctx context.Context, uid uint32, bm *roaring.Bitmap) error {
-	shard := int(uid / g.shardSize)
-	db, err := g.initShardDB(ctx, shard)
-	if err != nil {
-		return fmt.Errorf("failed to initialize shard db: %w", err)
-	}
-
-	bmBytes, err := bm.ToBytes()
-	if err != nil {
-		return fmt.Errorf("failed to marshal bitmap: %w", err)
-	}
-
-	if _, err := db.ExecContext(ctx, `INSERT OR REPLACE INTO entities (uid, compressed_bitmap) VALUES (?, ?);`, uid, bmBytes); err != nil {
-		return fmt.Errorf("failed to update entity: %w", err)
-	}
-
-	return nil
-}
-
+// UpdateUIDMapping explicitly updates the UID to string ID mapping in the meta database
+// This should only be called by an external process when bulk updating the mapping
+// To initialize a new UID, use GetUID with initOnMiss=true (which will call this function)
 func (g *Group) UpdateUIDMapping(ctx context.Context, uid uint32, stringID string) error {
-	if _, err := g.metaDB.ExecContext(ctx, `INSERT OR REPLACE INTO entity_map (uid, string_id) VALUES (?, ?);`, uid, stringID); err != nil {
-		return fmt.Errorf("failed to update uid mapping: %w", err)
-	}
-
-	return nil
-}
-
-// loadFromMetaDB loads the UID to String mappings from the meta database
-func (g *Group) loadFromMetaDB(ctx context.Context) error {
-	ctx, span := tracer.Start(ctx, "loadFromMetaDB")
-	defer span.End()
-
-	span.SetAttributes(attribute.String("group", g.name))
-
-	rows, err := g.metaDB.QueryContext(ctx, `SELECT uid, string_id FROM entity_map;`)
-	if err != nil {
-		return fmt.Errorf("failed to query entity map: %w", err)
-	}
-	defer rows.Close()
-
-	g.stuLk.Lock()
-	g.utsLk.Lock()
-	g.nextLk.Lock()
-	defer g.stuLk.Unlock()
-	defer g.utsLk.Unlock()
-	defer g.nextLk.Unlock()
-
-	var nextUID uint32
-
-	for rows.Next() {
-		var uid uint32
-		var stringID string
-		if err := rows.Scan(&uid, &stringID); err != nil {
-			return fmt.Errorf("failed to scan entity map: %w", err)
-		}
-
-		g.stu[stringID] = uid
-		g.uts[uid] = stringID
-		if uid >= nextUID {
-			nextUID = uid + 1
-		}
-	}
-
-	g.nextUID = nextUID
-
-	return nil
-}
-
-// initializes a MetaDB for the group at bitmapper_db_dir/group_name/meta.db
-func (g *Group) initMetaDB(ctx context.Context) (*sql.DB, error) {
-	metaDBPath := fmt.Sprintf("%s/%s", g.dbDir, MetaDBName)
-	// Open the database
-	db, err := sql.Open("sqlite3", metaDBPath)
-	if err != nil {
-		return nil, fmt.Errorf("failed to open database at %q: %w", metaDBPath, err)
-	}
-
-	// Set pragmas
-	if _, err := db.ExecContext(ctx, `PRAGMA journal_mode=WAL;`); err != nil {
-		return nil, fmt.Errorf("failed to set journal mode at %q: %w", metaDBPath, err)
-	}
-
-	// Set synchronous mode to off to increase write performance
-	// not super risky with WAL mode and redundant disks
-	if _, err := db.ExecContext(ctx, `PRAGMA synchronous=off;`); err != nil {
-		return nil, fmt.Errorf("failed to set synchronous mode: %w", err)
-	}
-
-	// Create the table if it doesn't exist
-	if _, err := db.ExecContext(ctx, `CREATE TABLE IF NOT EXISTS entity_map (
-		uid INTEGER PRIMARY KEY,
-		string_id TEXT NOT NULL
-	);`); err != nil {
-		return nil, fmt.Errorf("failed to create entity_map table: %w", err)
-	}
-
-	return db, nil
-}
-
-// initializes a ShardDB for the group at bitmapper_db_dir/group_name/shard_%d.db
-func (g *Group) initShardDB(ctx context.Context, shard int) (*sql.DB, error) {
-	db, ok := g.dbCache.Get(shard)
-	if ok {
-		return db, nil
-	}
-
-	dbPath := fmt.Sprintf("%s/"+ShardDBPattern, g.dbDir, shard)
-	// Open the database
-	db, err := sql.Open("sqlite3", dbPath)
-	if err != nil {
-		return nil, fmt.Errorf("failed to open database: %w", err)
-	}
-
-	// Set pragmas
-	if _, err := db.ExecContext(ctx, `PRAGMA journal_mode=WAL;`); err != nil {
-		return nil, fmt.Errorf("failed to set journal mode: %w", err)
-	}
-
-	if _, err := db.ExecContext(ctx, `PRAGMA synchronous=off;`); err != nil {
-		return nil, fmt.Errorf("failed to set synchronous mode: %w", err)
-	}
-
-	// Create the table if it doesn't exist
-	if _, err := db.ExecContext(ctx, `CREATE TABLE IF NOT EXISTS entities (
-		uid INTEGER PRIMARY KEY,
-		compressed_bitmap BLOB NOT NULL
-	);`); err != nil {
-		return nil, fmt.Errorf("failed to create entities table: %w", err)
-	}
-
-	g.dbCache.Add(shard, db)
-
-	return db, nil
+	return g.updateUIDMapping(ctx, uid, stringID)
 }
